[Vulkan] Move texture config to its own files.
